add function to create and delete users' folders

+++ app/routes.py
@@ -1,33 +1,26 @@
-from app import app
-from app import db
-from app.forms import RegisterForm, LoginForm
-from app.entity import User
-import os
-from flask import render_template, send_from_directory, flash, request, redirect, url_for
+from app import app, db
+from app.forms import LoginForm, RegisterForm
+from flask import flash, redirect, request, url_for, render_template, send_from_directory, abort
 from flask_login import current_user, login_user, logout_user
+from config import uploadsdir
+import os
+from app.models import User
+import shutil
 
-@app.route('/favicon.ico')
-def favicon():
-    return send_from_directory(os.path.join(app.root_path, 'static'), 
-                               'favicon.ico')
-
-@app.route("/")
-@app.route("/index")
-def index():
-    return render_template("index.html")
 
 @app.route("/singup", methods=["POST", "GET"])
-def register(): 
+def register():
     form = RegisterForm()
     if form.validate_on_submit():
-        user = User.query.filter_by(username=form.username.data).first()
-        if user:
-            flash('Login is taken', 'warning') #fix this flash
-            return redirect(url_for("index"))
+        if User.query.filter_by(username=form.username.data).first():
+            flash('Username is already taken. Please choose a different one.', 'danger')
+            return redirect(url_for('register'))
+        
         user = User(username=form.username.data)
         user.set_password(form.password.data)
         db.session.add(user)
         db.session.commit()
+        os.mkdir(os.path.join(uploadsdir, str(user.id)))
         flash(f'Successful Sign up! Your Id is {user.id}', 'success')
         login_user(user)
         return redirect(url_for("index"))
@@ -35,32 +28,54 @@ def register():
 
 @app.route("/login", methods=["POST", "GET"])
 def login():
+    if current_user.is_authenticated:
+        return redirect(url_for("index"))
+    
     form = LoginForm()
-    if form.validate_on_submit():
+    if form.validate_on_submit(): 
         user = User.query.filter_by(username=form.username.data).first()
+        
         if user is None or not user.check_password(form.password.data):
             flash('Incorrect login or password', 'warning')
             return redirect(url_for("login"))
+        
         if user.status != "ACTIVE":
             flash("Your account was deactivated or banned", 'danger')
             return redirect(url_for("login"))
+        
         login_user(user, remember=form.remember_me.data)
         flash('Seccussful Log In', 'success')
         return redirect(url_for("index"))
     return render_template('login.html', title="Sign in", form=form)
 
-@app.route("/lk")
-def account():
-    return render_template("base.html", title="Account") 
+@app.route("/logout")
+def logout():
+    logout_user()
+    flash('You have logged out', 'warning')
+    return redirect(url_for('index'))
+
+@app.route('/favicon.ico')
+def favicon():
+    return send_from_directory(os.path.join(app.root_path, 'static'), 'favicon.ico')
+
+@app.route("/")
+@app.route("/index")
+def index():
+    return render_template("index.html")
 
 @app.route("/admin", methods=["POST", "GET"])
 def admin():
+    if current_user.role != "ROOT":
+        abort(403)
+
     action = request.args.get("action", "")
     username = request.args.get("username")
     status = request.args.get("status")
     users = User.query.all()
+
     if action == "delete":
         user_to_delete = User.query.filter_by(username=username).first()
+        shutil.rmtree(os.path.join(uploadsdir, str(user_to_delete.id)), ignore_errors=True)
         db.session.delete(user_to_delete)
         db.session.commit()
         flash(f'User {user_to_delete.username} was successful deleted', 'success')
@@ -72,10 +87,10 @@ def admin():
         db.session.commit()
         return redirect("/admin?action=list_users")
 
-
     if action == "delete_db":
         for user in users:
             if user.role != "ROOT" or user.username != "root":
+                shutil.rmtree(os.path.join(uploadsdir, str(user.id)), ignore_errors=True)
                 db.session.delete(user)
                 db.session.commit()
         # for table_name in db.metadata.tables:
@@ -84,8 +99,6 @@ def admin():
         # return redirect(url_for("index"))
     return render_template("root.html", route=action, users=users)
 
-@app.route("/logout")
-def logout():
-    logout_user()
-    flash('You have logged out', 'warning')
-    return redirect(url_for('index'))
+@app.route("/account")
+def account():
+    return render_template("base.html", title="Account")
